+++ test/test_formatter.py
@@ -4,6 +4,7 @@ import struct
 import sys
 import tempfile
 import unittest
+import six
 from mock import Mock, patch
 from nose.tools import *    # pylint: disable=wildcard-import, unused-wildcard-import
 
@@ -83,9 +84,8 @@ class TestGetTerminalSize(unittest.TestCase):
 
 
 def _tf():
-    '''Open a temp file that looks a bunch like stdout.
-    '''
-    if sys.version_info[0] == 3:
+    """Open a temp file that looks a bunch like stdout."""
+    if six.PY3:
         # in python3 it's got an encoding and accepts new-style strings
         return tempfile.TemporaryFile(mode='w', encoding='UTF-8')